content: some minor edits on wording (#26)
* In the Stackdriver setup codelab, mention the APIs that we
are enabling i.e. "Stackdriver Tracing and Monitoring"

*  In content/quickstart/*, change the mention of "meta-data"
since "meta-data" while technically right and in fact the right
wording since they are optional and are meant to describe what
"tags" are about, OpenCensus users will subjectively think "tags"
are being relegated to a second class citizen

* In the tutorials, instead of saying a "REPL function", change
that say a "REPL app" because essentially the user is running
an app in its entirety.

* In content/quickstart/*/metrics.md, remove unused metrics:
"Number of spaces" and "Number of bytes read" as I didn't include
them in the actual code, hence users would perhaps get lost asking
for them. The ones we have seem sufficient.

* In content/quickstart/java/*, change the path from
"Put this in `/src/main/java/*`"
to
"Put this in `src/main/java/*`"
because we already assume the user in the current working directory
but if a user copies and pastes `/src/main/java/*`, they'll be
using a path relative to root and will cause confusion,
while `src/main/java/*` will use a path relative the current working
directory which is what we are aiming for.